+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.1.0
 
 - Enh #1: Introduce `Enum` helper class with tests (@terabytesoftw)
+- Bug #2: Update assertion messages for `Enum` class tests (@terabytesoftw)